refactor(ingesters): extract sead change request helpers
Closes #444
